requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
# Pillow==10.2.0
python-decouple==3.8
django-filter==23.5
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
python-docx==1.1.0
django-ratelimit==4.1.0
celery==5.3.4
redis==5.0.1
drf-yasg==1.21.7
setuptools>=65.5.0

